--- 1-todo/Counter.jsx
@@ -1,9 +1,0 @@
-
-import React from "react";
-
-const SS = () => {
-    return (
-        <div></div>
-    )
-
-}
