--- 1-todo/src/App.js
@@ -1,27 +1,0 @@
-import React, { Component } from 'react';
-import logo from './logo.svg';
-import './App.css';
-
-function test() {
-
-}
-
-class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <header className="  App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <h1 className="App-title">王跃林</h1>
-        </header>
-        <p className="App-intro">
-          To get started, edit <code>src/App.js</code> and save to reload.
-        </p>
-      </div>
-    );
-  }
-}
-
-
-
-export default App;
