+++ 1-todo/wangyuelin/Homework.css
@@ -1,0 +1,27 @@
+
+/* clear */
+.clearfix:after {
+    display:block;
+    clear:both;
+    content:"";
+    visibility:hidden;
+    height:0
+} 
+.clearfix {
+    zoom:1
+} 
+
+
+.row{
+    
+}
+.col {
+    display: inline-block;
+    float: left;
+}
+
+.checked {
+    border-bottom: 1px solid red;
+}
+    
+
